Started Prologue Rewrite!
The story is going quite well! The tale is now having Thorn be inspired by a historic account of a Thief named thorn which also tells a little about the lore of the world, and sets up curiosity for why Users went from heroes to monsters. The discovery of Ichor and having Thorn believe that it's blood from the gods is also a nice twist, like she's now living the story or discovering a mystery that could give her great strength. Next step, finish the back home part of the prologue, and then start on Chapter two!